add bridging tickets to mqtt broker
