+++ pyproject.toml
@@ -10,6 +10,7 @@ requests = "^2.27.1"
 argparse = "^1.4.0"
 python-telegram-bot = "^13.12"
 rich = "^13.4.1"
+paho-mqtt = "^1.6.1"
 
 [tool.poetry.dev-dependencies]
 black = "^22.3.0"
